TCP Added

+++ Updater/ViewModels/MainViewModel.cs
@@ -1,13 +1,16 @@
 ﻿using System;
 using System.Diagnostics;
+using System.Dynamic;
 using System.IO;
 using System.IO.Compression;
 using System.Reflection;
+using System.Runtime.InteropServices;
 using System.ServiceProcess;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Threading;
+using Microsoft.Win32;
 using Updater.API;
 using Updater.Core;
 
@@ -20,7 +23,22 @@ namespace Updater.ViewModels
         private int _percent;
         private string _title;
 
-        private string CurrentApplicationDir => Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
+        [Flags]
+        enum MoveFileFlags
+        {
+            MOVEFILE_REPLACE_EXISTING = 0x00000001,
+            MOVEFILE_COPY_ALLOWED = 0x00000002,
+            MOVEFILE_DELAY_UNTIL_REBOOT = 0x00000004,
+            MOVEFILE_WRITE_THROUGH = 0x00000008,
+            MOVEFILE_CREATE_HARDLINK = 0x00000010,
+            MOVEFILE_FAIL_IF_NOT_TRACKABLE = 0x00000020
+        }
+
+        [DllImport("kernel32.dll", SetLastError = true)]
+        static extern bool DeleteFile(string lpFileName);
+
+        [DllImport("kernel32.dll", SetLastError = true, CharSet = CharSet.Unicode)]
+        static extern bool MoveFileEx(string lpExistingFileName, string lpNewFileName, MoveFileFlags dwFlags);
 
         public string Title
         {
@@ -50,19 +68,18 @@ namespace Updater.ViewModels
            ContinueUpdating(fileName);
         }
 
-        #region MyRegion
-
         public async void StartUpdating()
         {
-            await Task.Run(() =>
+            await Task.Run(async () =>
             {
-                Title = "Closing application";
-                KillProcess(GlobalData.PROCESS_NAME);
-
-                Title = "Stopping services";
-                StopService();
-
                 Title = "Downloading";
+                await LogFileAsync("Stopping Service");
+                StopService();
+                
+                await LogFileAsync("Killing Process");
+                KillProcess();
+
+                await LogFileAsync("Start downloading");
                 _downloader = new FileDownloader();
                 _downloader.DownloadProgress += x => Percent = x;
                 _downloader.FileDownloadCompleted += FileDownloadCompleted;
@@ -70,33 +87,22 @@ namespace Updater.ViewModels
             });
         }
 
-        private async void ContinueUpdating(string fileName)
+        private async void ContinueUpdating(string filePath)
         {
             try
             {
-                await Task.Run(() =>
+                await Task.Run(async () =>
                 {
-                    var appData = Path.GetDirectoryName(fileName);
-                    var tempInstaller = Path.Combine(appData, GlobalData.FILE_INSTALLER);
-                    var tempShared = Path.Combine(appData, GlobalData.FILE_SHARED);
+                    Title = "Updating";
+                    var dirApp = GetInstallDirectory();
+                    await LogFileAsync("Start extracting");
+                    ExtractFilesAndReplace(filePath, dirApp);
+                    
+                    await LogFileAsync("Running");
+                    await RunAsync(dirApp, GlobalData.FILE_INSTALLER);
 
-                    var installerFile = Path.Combine(CurrentApplicationDir, GlobalData.FILE_INSTALLER);
-                    var sharedFile = Path.Combine(CurrentApplicationDir, GlobalData.FILE_SHARED);
-
-                    CleanUp(tempInstaller, tempShared, installerFile, sharedFile);
-
-                    ZipFile.ExtractToDirectory(fileName, appData);
-
-                    File.Copy(tempInstaller, installerFile);
-                    File.Copy(tempShared, sharedFile);
-
-                    Title = "Running application";
-                    Run(installerFile);
-
-                    Title = "Starting service";
+                    await LogFileAsync("Running Service");
                     StartService();
-
-                    Title = "Done";
                     Exit();
                 });
             }
@@ -106,7 +112,77 @@ namespace Updater.ViewModels
             }
         }
 
-        #endregion
+        private void ExtractFilesAndReplace(string filePath, string dirInstall)
+        {
+            var dirSrc = Path.GetDirectoryName(filePath);
+            var dirExtract = Path.Combine(dirSrc, "extract");
+            
+            MakeEmptyDirectory(dirExtract);
+            
+            ZipFile.ExtractToDirectory(filePath, dirExtract);
+            
+            var files = Directory.GetFiles(dirExtract);
+            foreach (var file in files)
+            {
+                var target = Path.Combine(dirInstall, Path.GetFileName(file));
+                MoveFileEx(file, target, MoveFileFlags.MOVEFILE_REPLACE_EXISTING);
+            }
+        }
+
+        private void MakeEmptyDirectory(string dir)
+        {
+            if (Directory.Exists(dir))
+            {
+                foreach (var file in Directory.GetFiles(dir))
+                    DeleteFile(file);
+            }
+            else
+                Directory.CreateDirectory(dir);
+        }
+
+       private async void MoveFile(string dirSrc, string dirTarget, string fileName)
+        {
+            var src = Path.Combine(dirSrc, fileName);
+            var target = Path.Combine(dirTarget, fileName);
+            
+            await LogFileAsync($"Deleting {target}");
+            DeleteFile(target);
+            
+            await LogFileAsync($"Moving from: '{src}' to '{target}'");
+            MoveFileEx(src, target, MoveFileFlags.MOVEFILE_DELAY_UNTIL_REBOOT);
+        }
+
+        private string GetInstallDirectory()
+        {
+            var reg = Registry.CurrentUser.CreateSubKey(@"SOFTWARE\InstallerService");
+            return (string) reg.GetValue("ApplicationDirectory", "");
+        }
+
+        private void DeleteFile(string dir, string fileName)
+        {
+            var filePath = Path.Combine(dir, fileName);
+            DeleteFile(filePath);
+        }
+
+        //private bool DeleteFile(string filePath)
+        //{
+        //    for (int i = 0; i < 10; i++)
+        //    {
+        //        try
+        //        {
+        //            if (File.Exists(filePath))
+        //                File.Delete(filePath);
+        //            return true;
+        //        }
+        //        catch (Exception)
+        //        {
+        //        }
+
+        //        Thread.Sleep(200);
+        //    }
+
+        //    return false;
+        //}
 
         #region system files
 
@@ -123,7 +199,7 @@ namespace Updater.ViewModels
                         TimeSpan.FromMilliseconds(GlobalData.SERVICE_TIMEOUT));
                 }
             }
-            catch(Exception ex)
+            catch (Exception ex)
             {
                 MessageBox.Show(ex.Message);
             }
@@ -142,33 +218,42 @@ namespace Updater.ViewModels
                         TimeSpan.FromMilliseconds(GlobalData.SERVICE_TIMEOUT));
                 }
             }
-            catch(Exception ex)
+            catch (Exception ex)
             {
                 MessageBox.Show(ex.Message);
             }
         }
 
-        public static void KillProcess(string processName = GlobalData.PROCESS_NAME)
+        public async void KillProcess(string processName = GlobalData.PROCESS_NAME)
         {
-            var processes = Process.GetProcessesByName(processName);
-            foreach (var process in processes)
-                process.Kill();
+            try
+            {
+                var processes = Process.GetProcesses();
+                foreach (var process in processes)
+                {
+                    if (process.ProcessName.Contains(processName))
+                        process.Kill();
+                }
+            }
+            catch (Exception ex)
+            {
+                await LogFileAsync(ex.Message);
+            }
         }
 
-        private void Run(String fileName)
+        private async Task RunAsync(string dir, String fileName)
         {
-            if (File.Exists(fileName) == false)
+            var file = Path.Combine(dir, fileName);
+            if (File.Exists(file) == false)
+            {
+                await LogFileAsync($"Running {file}");
                 return;
+            }
 
-            var process = new Process {StartInfo = {FileName = fileName, Arguments = ""}};
+            await LogFileAsync($"Running {file}");
+            var process = new Process { StartInfo = { FileName = file, Arguments = "" } };
             process.Start();
-        }
-
-        private void CleanUp(params string[] files)
-        {
-            foreach (var file in files)
-                if (File.Exists(file))
-                    File.Delete(file);
+            return;
         }
 
         private async void Exit()
@@ -178,6 +263,14 @@ namespace Updater.ViewModels
         }
 
         #endregion
+
+        private async Task LogFileAsync(string message)
+        {
+            using (var file = File.AppendText("d:\\logTest.txt"))
+            {
+                await file.WriteLineAsync(message + "\n");
+            }
+        }
 
         public MainViewModel()
         {
